Struct bytestring::ByteString

pub struct ByteString(_);

An immutable UTF-8 encoded string with Bytes as a storage.

Implementations

impl ByteString

pub const fn new() -> Self

Creates a new empty ByteString.

pub fn as_bytes(&self) -> &Bytes

Get a reference to the underlying Bytes object.

pub fn into_bytes(self) -> Bytes

Unwraps this ByteString into the underlying Bytes object.

pub const fn from_static(src: &'static str) -> ByteString

Creates a new ByteString from a &'static str.

pub const unsafe fn from_bytes_unchecked(src: Bytes) -> ByteString

Creates a new ByteString from a Bytes.

Safety

This function is unsafe because it does not check the bytes passed to it are valid UTF-8. If this constraint is violated, it may cause memory unsafety issues with future users of the ByteString, as we assume that ByteStrings are valid UTF-8. However, the most likely issue is that the data gets corrupted.